Refactor classes separate packages
