+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## v1.1.0 (TBD)
+
+- Refactor classes separate packages
+
 ## v1.0.0 (2023-04-30)
 
 - Initial release using https://api.dailytsp.com/close/, based on scholarshare.price
