+++ src/main/java/tsp/price/Application.java
@@ -7,6 +7,7 @@ import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Scope;
 import org.springframework.web.client.RestTemplate;
+import tsp.price.xfer.Request;
 
 /**
  * Downloads TSP share price information
