+++ src/main/java/tsp/price/data/Observations.java
@@ -1,4 +1,4 @@
-package tsp.price;
+package tsp.price.data;
 
 import lombok.Builder;
 import lombok.Data;
